Prevented unnecessary copying in INI parser
The INI parser used to use `std::string` as opposed to `std::string&`
in most function parameters, which comes at the cost of making a full
string copy. Since we're not modifying strings in-place, this can be
skipped.

Further, this commit removes those overloads that specify unused flags,
as this gets `PDINI.hpp` through compilations with:

-pedantic -Wall -Wextra -Werror

A long-term goal would be to enforce these compilation flags in
`CMakeLists.txt`, but more clean-up work would be necessary for that.