Separated third-party code from first-party code
This commit moves

- the currently unused PCG library,
- the to-be-externalized PDINI library,
- the generated `glad` loader

to `third-party`, a new directory.

This allows us to set the third-party directory as a system inclusion
directory in CMake, resulting in gcc and clang using `-isystem` as
opposed to `-I`. The reason why we're doing that is to prevent us from
touching third-party code to fix errors we would otherwise see when
using `-pedantic` and related flags.