Removed `using namespace` pattern
`using namespace std;` is a notoriously bad anti-pattern in C++.
The same argument applies to other namespaces, too: We should try our
best to prevent name clashes, so let's be explicit about the namespaces
our definitions are put in.

For now, there are two solutions to refer to the `Shiro` namespace:

(1) namespace Shiro { /* contents of C++ source */  }

This prevents us from accidentally polluting the global namespace and
is to preferred over (2).

(2) Shiro::*

This is used when (1) wasn't possible. The reason for it not being
possible was that the code currently relies on global namespace
pollution, as some function definitions are declared/defined there.

These instances of (2) will be converted to (1) once we abandon the
C-style remnants in the code base.