Moved build options to CMake `option`, enabled interpolation by default
This should really be an opt-out feature, since most machines will very
likely match the necessary requirements. In order to make the build
process less annoying on Apple machines, the option will still be
disabled by default in case OpenGL is not available on the machine.

Strategically, we should try to ship the best product by default, which
becomes especially important for automated builds. There is barely a
reason to force users to build their own custom executable, just to get
access to hidden gems.

This commit also reduces redundance and superfluous lines in our CMake
build file that aren't needed to get the project to compile.