Optimized audio assets
In an attempt to lessen the impact of our dynamic asset header
generation to compile-time memory requirements, this commit optimizes
all audio assets by converting them to Ogg vorbis.

This dramatically reduces the dynamic audio header size from around
100 MiB to just 12.4 MiB, which corresponds to an improvement of
approximately 833 %.

The bitrate is different for every file, since I have invested a
considerable amount of time going through each one individually and
finding the sweet spot where no change in quality is noticeable
anymore.

The reason why this work is done is to make it possible to fix the
issue with black screens where the game can't find the asset folder,
as was described in #22.