Linked against the `WINDOWS` subsystem in Windows Release builds
On Windows, the game's release build should be linked against the
WINDOWS subsystem as opposed to the `CONSOLE` one, which it has been
prior to this commit.

Further, we now link release builds with `/ENTRY:mainCRTStartup`,
which causes `main` to be the entry point on Windows.

In essence, this commit prevents the console window used for debugging
to show in release builds on Windows.