Cleaned up `CMakeLists.txt`
This commit makes the CMake file a little cleaner, such that it can be
more easily maintained. In particular, the executable target is now
defined without sources and headers, as they are added piece-by-piece
depending on the configuration-time definitions.

Next, IDE `source_group`s are generated from configuration-time sources
and now include the "Third-party" group.

This commit also enables automatic memory sanitization on gcc and clang
when built in `Debug` mode.

Also, the project version has been changed to `0.1.0` for now. This
is a step towards solving #11, though re-tagging old tags still remains
to be done.

Further experimentation is required to see if we could eliminate our
`PkgConfig` package dependency for those modules that CMake can find on
its own.