Implemented means to start the game without a CWD lookup
Looking up a `game.ini` in the current working directory is confusing
if you navigate with a terminal, as is common in Linux.

If you were lucky, you remembered to change your directory into that
of the executable before running it, or maybe you were lucky enough to
be in the repository's root directory that also includes a `game.ini`.

With this commit, you can be in any directory, and the configuration
lookup instead takes place wherever the game executable resides.

This commit fixes another instance of the behavior described in #22.