Fixed bug in `BASE_PATH` resolution, removed positional arguments
Prior to this commit, a relative `BASE_PATH` would be assumed to be
relative to your current working directory. With this commit, it is
assumed to be relative to the directory that your configuration file
is stored at.

Further, in order to make the game more more configurable from the
command line, I'm migrating the positional configuration-path argument
to a flag-based argument, putting it behind `--configuration-path` or
`-c`.

This is a tiny step towards #45.