Configured `VS_DEBUGGER_COMMAND_ARGUMENTS` to specify `game.ini`
Visual Studio has no way of knowing where the assets are stored.
Therefore, the game can't be run without setting up a launch
configuration for the debugger.

With CMake, we can specify a default launch configuration that the
debugger uses. The goal here is to have Windows users know as little
as possible in order to get the game to compile and run.

This commit also renames

	`--configuration-path`

to

	`--configuration-file`.