Removed `sectionName` parameter for `keyBindings`, `sectionBindings`
They are always the same, so it doesn't make sense to parameterize
them right now. In addition, I'm also starting to see some of the
downsides of the ini file format.

For one, it's unclear whether users need to quote RHS string values or
not. Unquoted paths with spaces feel wonky.

Most importantly though, as the game grows, the sections will likely
become pretty unimportant, as 100 or more options for the input section
alone are realistic. All in all, maybe we should just settle with a
format that maps

key = value

and nothing more than that. Then, namespacing becomes possible through
a key prefix alone. This is also what RetroArch does.