Fixed most warnings when compiling with Visual Studio 2019 all the way up to warning level 4 and with conformance mode on (progress on #24).
There are many warnings that I didn't bother with fixing, because they're either just suggested changes that we don't need to bother with, such as using unscoped enums as opposed to enum class (which can't be fully fixed anyways, because those are used in the C library headers, such as SDL.h), or false positives due to the compiler not being able to figure out there's nothing wrong. There are also some warnings that can't be easily fixed, due to the offending code being very C-style, and Visual Studio not properly analyzing said code.

More testing with other compilers/platforms needs to be done to be sure no more warnings worth fixing remain.