Made the game able to start without `shiromino.ini` (closes #20).
When no `shiromino.ini` is present, defaults are selected, and the configuration is written to a new `shiromino.ini`, at the location where it would have loaded from if it were already present.

For now, the already-present prewritten `shiromino.ini` will remain in the repo, because it documents the options so users can know how to manually edit it. When full in-game configuration is in place, that configuration file can be removed, and we can rely on the default configuration code for new installations.