Removed case that would make garbage creation fail
This is an attempt to fix garbage. According to Felicity, that case
was planned to be used for fixed garbage shapes.

Since all it contains is a `// TODO`, I'm assuming that this hasn't
been implemented yet.

However, it seems that `q->garbage` can still contain a `Shiro::Grid`
of size 0 sometimes, making garbage creation fail.