Restructured things a bit
This commit ports all SGUIL macros to `constexpr` expressions.

Further, this commit renames `SGUIL` to just `GUI`, as it is more
expressive about what it's supposed to mean (and it is also easier to
pronounce).

Lastly, this commit removes a few unnecessary files, unnecessary code,
unnecessary macros, and renames `Gui` to `GUI`.