Added missing dates to `CHANGELOG.md`
These are needed to make it easier for us to determine when exactly
our versions have been released to the public.

Our git history is subject to change, so tag dates may change, and
our presence on GitHub should be considered ephemeral, meaning that
the release page is not a valid source of truth.

Moreover, GitHub resets the release date to the tag's COMMIT date
instead of the AUTHOR date, which is completely insane, makes
everything unreadable and wrong, and should be prosecuted.