Improve performance on and support for Windows.
An improved "Delay" function that replaces SDL_Delay when on Windows, used when building with MinGW GCC/Clang or MSVC, is provided, but delays are done with SDL_Delay on all other platforms; SDL_Delay implements delays with nanosleep on Unix platforms, which performs well, but SDL_Delay uses Sleep on Windows, which produces inconsistent and often-too-long delays, vs. the technique now used in Delay on Windows. And building with MSVC/Visual Studio is fully supported again, with some cleanups that make the support not require any "ugly" changes to accommodate MSVC.